Change the API a bit. Use values between 0 and 1.

+++ stepmania/src/archutils/Darwin/DarwinThreadHelpers.h
@@ -4,12 +4,13 @@
 bool SuspendThread( uint64_t threadHandle );
 bool ResumeThread( uint64_t threadHandle );
 uint64_t GetCurrentThreadId();
-void SetThreadPrecedence( int prec );
+// Valid values are from 0.0f to 1.0f. 0.5f is default.
+void SetThreadPrecedence( float prec );
 
 #endif
 
 /*
- * (c) 2004, 2005 Steve Checkoway
+ * (c) 2004-2006 Steve Checkoway
  * All rights reserved.
  * 
  * Permission is hereby granted, free of charge, to any person obtaining a
